+++ service/order/dto/src/main/java/dto/OrderDto.java
@@ -4,14 +4,9 @@ import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Getter
-@AllArgsConstructor
-@Builder
-@NoArgsConstructor
 public class OrderDto {
 
   @Getter
