+++ service/order/server/docs/OrderApiTest.http
@@ -9,7 +9,7 @@ Content-Type: application/json
   "orderType": "STANDARD",
   "orderProductInfos": [
     {
-      "productId": "19646241-55d8-40f7-8f3e-d7242dedff49",
+      "productId": "800504d0-be21-450e-8598-0dc3d382d7ea",
       "quantity": 3,
       "userCouponId": 132
     }
@@ -18,6 +18,10 @@ Content-Type: application/json
   "addressId": 1
 }
 
+### 주문 취소
+PATCH http://localhost:{{Port}}/api/orders/19149/cancel
+Authorization: Bearer {{access_token}}
+
 ### 상품 재고 조회
-GET http://localhost:{{Port}}/api/products/19646241-55d8-40f7-8f3e-d7242dedff49
+GET http://localhost:{{Port}}/api/products/800504d0-be21-450e-8598-0dc3d382d7ea
 Authorization: Bearer {{access_token}}
