+++ service/order/server/src/main/java/com/sparta/order/server/presentation/controller/OrderController.java
@@ -2,10 +2,13 @@ package com.sparta.order.server.presentation.controller;
 
 import com.sparta.auth.auth_dto.jwt.JwtClaim;
 import com.sparta.common.domain.response.ApiResponse;
+import com.sparta.order.server.application.service.OrderCreateService;
 import com.sparta.order.server.application.service.OrderService;
 import com.sparta.order.server.presentation.dto.OrderDto.OrderCreateRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,12 +19,19 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class OrderController {
 
+  private final OrderCreateService orderCreateService;
   private final OrderService orderService;
 
   @PostMapping
   public ApiResponse<Long> createOrder(@AuthenticationPrincipal JwtClaim userClaim,
       @RequestBody OrderCreateRequest request) {
-    return ApiResponse.created(orderService.createOrder(userClaim.getUserId(), request));
+    return ApiResponse.created(orderCreateService.createOrder(userClaim.getUserId(), request));
+  }
+
+  @PatchMapping("/{orderId}/cancel")
+  public ApiResponse<Long> cancelOrder(@AuthenticationPrincipal JwtClaim userClaim,
+      @PathVariable(name = "orderId") Long orderId) {
+    return ApiResponse.ok(orderService.cancelOrder(userClaim.getUserId(), orderId));
   }
 
 }
