+++ service/order/server/src/main/resources/application-local.yml
@@ -14,7 +14,7 @@ spring:
 
   data:
     redis:
-      host: cart-cache
+      host: localhost
       port: 6379
       username: default
       password: systempass
